wren_agent/pyproject.toml
[project]
name = "wren-agent"
version = "0.1.0"
description = "AI agent for generating Wren SDK scripts"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "openai-agents>=0.7.0",
    "pydantic>=2.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
    "python-dotenv>=1.0.0",
    "semgrep>=1.50.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
wren-agent = "agent.main:main"

# [tool.uv.sources]
# openai-agents = { path = "../examples/openai-agents-python" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
